chore(api): expose storage path environment variable
The storage path configuration has been moved from a hardcoded value to an environment variable configuration. This allows more flexibility in defining the storage location for the storage module. The `STORAGE_PATH` environment variable is now part of the `ENV_CONFIG_SCHEMA` with a default value of '.data/storage' and is used when creating the default storage pool.